Updated IncludeAssetLibrary and AssetThemeConfigurationServlet to support theme configurations
Updated IncludeAssetLibrary to accept a themeConfigPath that takes precedence over a simple theme include. This will
render the css path to be theme's path location + the theme configuration's unique name (as a selector).

Updated the AssetThemeServlet to render the theme's configuration if the selector points to a valid theme configuration
(by name). This is done so that relative paths from the theme css still work. Removed AssetThemeConfigurationServlet as
this caused the relative path syndrome just mentioned.