Adding nodetypes.cnd to air-ui package to register cystin namespaces during package install
The namespace is installed on bundle startup, but AEM package manager may not start the air service bundle before it
starts installing any other downward dependencies that may depend on the namespace to be registered. Bt defining the
custom namespace in a nodetypes.cnd this forces the namespace to be installed on the UI package install.